refactor to reuse code, clean up, add subscription
